RUN-3175: renamed webrequest_handler.ts to web_request_handler.ts
