+++ index.js
@@ -43,7 +43,7 @@ let socketServer = require('./src/browser/transports/socket_server').server;
 let authenticationDelegate = require('./src/browser/authentication_delegate.js');
 let convertOptions = require('./src/browser/convert_options.js');
 let coreState = require('./src/browser/core_state.js');
-let webRequestHandlers = require('./src/browser/webrequest_handler.js');
+let webRequestHandlers = require('./src/browser/web_request_handler.js');
 let errors = require('./src/common/errors.js');
 import ofEvents from './src/browser/of_events';
 import {
