+++ src/browser/web_request_handler.ts
